+++ pymc3/step_methods/particle.py
@@ -1,5 +1,7 @@
+import numpy as np
 import pymc3 as pm
 import theano
+from pymc3 import Point
 
 from .arraystep import BlockedStep
 from ..blocking import ArrayOrdering, DictToArrayBijection
@@ -59,3 +61,49 @@ def logp(logp, vars, shared):
     f = theano.function([inarray0], logp0)
     f.trust_input = True
     return f
+
+
+def transform_start_particles(start, nparticles, model=None):
+    """
+    Parameters
+    ----------
+    start : list, dict
+        Accepts 3 data types:
+        * dict of start positions with variable shape of (nparticles, var.dshape)
+        * dict of start positions with variable shape (var.dshape)
+        * list of dicts of length nparticles each with start positions of shape (var.dshape)
+    nparticles : int
+    model : Model (optional if in `with` context)
+
+    Returns
+    -------
+     transformed_start : a dict each with start positions of shape (nparticles, var.dshape)
+    """
+    if start is None:
+        return {}
+    dshapes = {i.name: i.dshape for i in model.vars}
+    dshapes.update({i.name: i.transformed.dshape for i in model.deterministics if hasattr(i, 'transformed')})
+    if isinstance(start, dict):
+        start = {k: v for k,v in start.items() if k in dshapes}
+        if all(dshapes[k] == np.asarray(v).shape for k, v in start.items()):
+            if nparticles is not None:
+                start = {k: np.asarray([v]*nparticles) for k, v in start.items()}  # duplicate
+            return Point(**start)
+        else:
+            extra = tuple() if nparticles is None else (nparticles, )
+            if all(extra+dshapes[k] == np.asarray(v).shape for k, v in start.items()):
+                return Point(**start)
+            else:
+                raise TypeError("Start dicts must have a shape of (nparticles, dshape) or (dshape,)")
+    elif isinstance(start, list):
+        start = [{k: v for k, v in s.items() if k in dshapes} for s in start]
+        assert len(start) == nparticles, "If start is a list, it must have a length of nparticles"
+        assert all(isinstance(i, dict) for i in start), "Start dicts must have a shape of (dshape,)"
+        assert all(s.keys() == start[0].keys() for s in start), "All start positions must have the same variables"
+        d = {}
+        for varname, varshape in dshapes.items():
+            if varname in start[0].keys():
+                assert all(varshape == s[varname].shape for s in start), "Start dicts must have a shape of (dshape,)"
+                d[varname] = np.asarray([s[varname] for s in start])
+        return Point(**d)
+    raise TypeError("Start must be a dict or a list of dicts")
